fix(automod): moderates edited messages too

+++ src/events/messageCreate.ts
@@ -7,34 +7,12 @@ import { Bot } from "../structures/Bot";
 import { TypedEvent } from "../types/types";
 import utils from "./../utils";
 
-const forbiddenPhrases: string[] = ["porn", "orange youtube", "faggot", "kys"];
-
 export default TypedEvent({
     eventName: "messageCreate",
     run: async (client: Bot, message: Message) => {
         if (!message.content || message.author.bot) return;
 
-        const messageWords = weirdToNormalChars(
-            message.content.toLowerCase()
-        ).split(" ");
-        const foundPhrase = forbiddenPhrases.some(
-            (phrase) =>
-                messageWords.join(" ").includes(phrase.toLowerCase()) ||
-                stringSimilarity(messageWords.join(" "), phrase) > 0.7 ||
-                messageWords.some(
-                    (word) => stringSimilarity(word, phrase) > 0.7
-                )
-        );
-        if (foundPhrase) return message.delete();
-
-        const inviteURLs = message.content.match(Invite.INVITES_PATTERN) ?? [];
-        for (const inviteURL of inviteURLs) {
-            const invite = await client
-                .fetchInvite(inviteURL)
-                .catch(() => null);
-            if (invite && invite.guild?.id !== config.guildId)
-                return await message.delete();
-        }
+        if (await utils.badContent(message)) return await message.delete();
 
         if (
             message.mentions.users.size > 5 &&
